+++ wefact/enums/__init__.py
@@ -1,3 +1,4 @@
 from .actions import Action
 from .invoice_actions import InvoiceAction
 from .interaction_actions import InteractionAction
+from .cost_category_actions import CostCategoryAction