+++ wefact/enums/actions.py
@@ -6,8 +6,3 @@ class Action(str, Enum):
     ADD = 'add'
     EDIT = 'edit'
     DELETE = 'delete'
-    COSTCATEGORY_LIST = 'costcategory_list'
-    COSTCATEGORY_SHOW = 'costcategory_show'
-    COSTCATEGORY_ADD = 'costcategory_add'
-    COSTCATEGORY_EDIT = 'costcategory_edit'
-    COSTCATEGORY_DELETE = 'costcategory_delete'