+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.1] - 2025-08-10
+
+**Changed**
+
+- Moved cost category-specific action constants out of the generic `Action` enum.
+	- New enum: `CostCategoryAction` in `wefact.enums.cost_category_actions` (also exported via `wefact.enums`).
+	- Replace usages of `Action.COSTCATEGORY_*` with `CostCategoryAction.*`.
+
 ## [1.1.0] - 2025-08-10
 
 **Added**